Manual merged #43 Mac OS Compilation fixes

+++ src/Mem_unix.cpp
@@ -47,7 +47,7 @@ bool Mem::allocate(const Options* options)
     m_memorySize = 0;
 
     size_t scratchPadSize = m_algo == Options::ALGO_CRYPTONIGHT ? MEMORY : MEMORY_LITE;
-    for (int i=0; i < m_threads; i++) {
+    for (size_t i=0; i < m_threads; i++) {
         m_memorySize += sizeof(cryptonight_ctx);
         m_memorySize += scratchPadSize * getThreadHashFactor(i);
     }
